Optimize prefetch function
Previously the prefetch function would have made a lot of postconf
calls, since it called fetch_resources for every resource in the
catalog. This commit limits the calls to every distinct confdir, which
can have a large impact on the runtime (in case there are lot of resources
of that type). Thanks to Raphaël Pinson who brought me on this and helped
with an efficient implementation.