Bugfix: Actually associate existing resources with each other
Previously the provider was always trying to create ressources although they
already existied. This was because the association of discovered resources
to catalog resources were not working properly. This was a) because of not
actually looping over the catalog resources and b) comparing the name
of one resource with an undefined name and therefore not matching
resources at all.